Take ownership of ndarray data inputs to prevent lifetime issues
Unfortunately, nvidia's "docs" (inline header comments) are vague and inaccurate. From examples, it appears that remarks like "function takes ownership of X" don't mean that the function will (ultimately) release X
but that X must not be altered or released before the function (or its object) is done with it.

https://github.com/triton-inference-server/core/blob/main/include/triton/core/tritonserver.h